all hashes added

+++ headers/hash.h
@@ -1,20 +1,24 @@
 //-------------------------------------------------------------\\
 //     Seven types of hashes are presented in this file        \\
+//-------------------------------------------------------------\\
+
+#include <string.h>
+#include <stdint.h>
 
 const int ERROR_HASH = 1e9 + 17;
 
 
-int hash_1(const char* inputString);
+uint64_t hash_1(const char* inputString);
 
-int hash_2(const char* inputString);
+uint64_t hash_2(const char* inputString);
 
-int hash_3(const char* inputString);
+uint64_t hash_3(const char* inputString);
 
-int hash_4(const char* inputString);
+uint64_t hash_4(const char* inputString);
 
-int hash_5(const char* inputString);
+uint64_t hash_5(const char* inputString);
 
-int hash_6(const char* inputString);
+uint64_t hash_6(const char* inputString);
 
-int hash_7(const char* inputString);
+uint64_t hash_7(const char* inputString);
